+++ src/tree/bst/waaforest/mod.rs
@@ -3,19 +3,19 @@
 //!
 use slab;
 use std::ops::{Index, IndexMut};
-use tree::bst::{BalancedBinaryForest, BstDirection, OrderedTree};
-use tree::traversal::{BinaryPreOrderIndices, Traversable};
-use tree::WeightedTree;
-use types::{DefaultWeightType, NodeIndex, Tgf, Values, ValueType, WeightType};
+use crate::tree::bst::{BalancedBinaryForest, BstDirection, OrderedTree};
+use crate::tree::traversal::{BinaryPreOrderIndices, Traversable};
+use crate::tree::WeightedTree;
+use crate::types::{DefaultWeightType, NodeIndex, Tgf, Values, ValueType, WeightType};
 
 #[cfg(test)]
 mod tests;
 
 #[derive(Clone, Debug)]
 struct Node<V, W>
-where
-    V: ValueType,
-    W: WeightType,
+    where
+        V: ValueType,
+        W: WeightType,
 {
     value: V,
     level: usize,
@@ -381,7 +381,7 @@ where
             if parent == self.nil {
                 break;
             }
-            path |= (dir as u64) << i;
+            path |= (dir as u64) << (i as u64);
             child = parent;
             parent = self.arena[child].parent;
             dir = if self.arena[parent][BstDirection::Left] == child {
